Moved legacy implementation of `Icon::ResolveFileExtensionIcon()` to `LegacyIconResolutionService::ResolveFileExtensionIcon(Icon& icon)`.
